#29: Restructure it even more.
