+++ src/main/java/org/weathertrack/validation/validator/InputValidator.java
@@ -1,4 +1,6 @@
-package org.weathertrack.validation;
+package org.weathertrack.validation.validator;
+
+import org.weathertrack.validation.model.InputValidationResult;
 
 public interface InputValidator {
 	InputValidationResult validateCityNameInput(String userCityInput);
