--- src/main/java/org/weathertrack/weather/WeatherDataProvider.java
@@ -1,5 +1,0 @@
-package org.weathertrack.weather;
-
-public interface WeatherDataProvider {
-	WeatherData getWeatherData(String city);
-}
