+++ src/main/java/org/weathertrack/weather/provider/WeatherDataProvider.java
@@ -1,0 +1,7 @@
+package org.weathertrack.weather.provider;
+
+import org.weathertrack.weather.model.WeatherData;
+
+public interface WeatherDataProvider {
+	WeatherData getWeatherData(String city);
+}
