+++ src/test/java/org/weathertrack/validation/validator/UserInputValidatorTests.java
@@ -1,10 +1,12 @@
-package org.weathertrack.validation;
+package org.weathertrack.validation.validator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.weathertrack.validation.exception.InputExceptionMessage;
+import org.weathertrack.validation.resource.ValidationMessageResource;
 
 import java.util.stream.Stream;
